GHA Deployment Automation
Why these changes are being introduced:
Application development has proceeded far enough alongt that it is time
to provide the Makefile commands for manual builds/deploys to Dev and
for automated dev/stage/prod deploys via GitHub.

How this addresses that need:
* Add Terraform-generated Makefile additions to make it easy for the
developer to build a container and push it to the correct ECR
repository in Dev1
* Create the dev, stage, and prod GitHub Actions to automatically
build and deploy the container to ECR and update the associated Lambda
function with the new image

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1267