Makefile and Dev Build Workflow
Why these changes are being introduced:
This is the initial Terraform-generated Makefile commands and dev-build
GitHub Actions workflow.

How this addresses that need:
* Add Terraform-generated Makefile additions to make it easy for the
developer to build a container and push it to the correct ECR
repository in Dev1
* Add Terraform-generated dev-build.yml workflow so that PRs to main
will automatically push a container to Dev1 and will automatically
deploy the updated Lambda function

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1218