More explicit with download and upload methods and args
Why these changes are being introduced:

Some PR feedback shared that it was a bit confusing how the utility
function stream_file_transfer() was capable of reading and writing
from a "URI" which could be an S3 object or a local filesystem path,
while the download and upload methods on the BagitArchive class seemed
more opinionated in what type of location was getting used.

There is some tension where a URI feels more accurate for the more
input/output agnostic utility function stream_file_transfer(), but
the methods are fairly opinionated whether things are coming from
S3 or not.

How this addresses that need:

Attempted to keep some flexibility in the utility function while making
the methods more opinionated by reworking some method names and arguments:

- upload_file() renamed to upload_bag_to_s3() which is quite literally
the only way it's used

- arguments for download_file() and upload_bag_to_s3() indicate if
they are "remote URIs" or some kind of "local path" to the degree possible

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1214