Configurable root workspace directory
Why these changes are being introduced:

This application needs to be able to create Bagit zip files that
are quite large.  As such, the AWS Lambda has been created with an
EFS mount for use.  To use this EFS mount, the application needs
to be configurable for where its "workspace" is for:

1. creating the empty Bag folder
2. downloading files to for the Bag
3. zipping up the Bag

It has been demonstrated this is memory safe, but it will require
actual filesystem space equal to the size of the bag.

How this addresses that need:
* Adds env var 'WORKSPACE_ROOT_DIR' which defines the root
working directory where *temporary* directories are created
for each Bag created
* Uses this workspace root when creating temporary Bagit
directories and files

Side effects of this change:
* When the env var is set, the Lambda will begin to use the
EFS mount at, at time of this writing, '/mnt/efs'.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1230