Bag zip as sibling to bag directory
Why these changes are being introduced:

A bug was discovered that after zipping up the bag,
there was a corupt, empty 'bag.zip' file in the downloaded
zip file.  This was suspiciously the name of the temporary
file used before upload.

How this addresses that need:

A little more care is taken when setting up temporary paths
for a) building the bag and b) creating the zip *from* the bag.

Confirmed that the final bag zip file does not contain an
empty 'bag.zip' in the contents.

Side effects of this change:
* None

Relevant ticket(s):
* None